object_recognition_system/requirements.txt
flask==1.1.2
jinja2==2.11.2
numpy==1.19.1
pillow==7.2.0